heap: use gsize for index_ parameter
This matches the length field more correctly, even though you wouldn't
want to do this.

+++ contrib/egg/egg-heap.c
@@ -316,7 +316,7 @@ egg_heap_extract (EggHeap  *heap,
 
 gboolean
 egg_heap_extract_index (EggHeap  *heap,
-                        guint     index_,
+                        gsize     index_,
                         gpointer  result)
 {
   EggHeapReal *real = (EggHeapReal *)heap;
